refactor : Int+Extension 파일 생성 및 ItemListViewController의 메서드 위치 수정
