+++ Expo1900/Expo1900/Model/DataManager.swift
@@ -38,6 +38,5 @@ struct DataManager {
         
         return dataList
     }
-    
 }
